+++ packages/cysync-core/src/pages/MainApp/Account/index.tsx
@@ -20,10 +20,16 @@ import {
   openEnableMergeDelegationDialog,
   openReceiveDialog,
   openSendDialog,
+  openSyncCantonAccountPromptDialog,
 } from '~/actions';
 import { CoinIcon, Graph, TransactionTable } from '~/components';
-import { constants } from '~/constants';
-import { selectLanguage, useAppDispatch, useAppSelector } from '~/store';
+import {
+  selectCantonAuthTokens,
+  selectCantonUnauthorizedSyncError,
+  selectLanguage,
+  useAppDispatch,
+  useAppSelector,
+} from '~/store';
 
 import { TokenTable } from './TokenTable';
 
@@ -41,6 +47,10 @@ const SendIcon = (props: SvgProps) => (
 export const AccountPage: FC = () => {
   const dispatch = useAppDispatch();
   const lang = useAppSelector(selectLanguage);
+  const cantonAuthTokens = useAppSelector(selectCantonAuthTokens);
+  const cantonUnauthorizedSyncError = useAppSelector(
+    selectCantonUnauthorizedSyncError,
+  );
 
   const {
     selectedWallet,
@@ -103,27 +113,10 @@ export const AccountPage: FC = () => {
               {lang.strings.buttons.receive}
             </Button>
             {selectedAccount?.familyId === coinFamiliesMap.canton &&
-              selectedAccount.type !== AccountTypeMap.subAccount && (
-                <Button variant="primary">
-                  <a
-                    href={constants.inheritance.cantonLink}
-                    target="_blank"
-                    style={{
-                      textDecoration: 'none',
-                      color: 'black',
-                      whiteSpace: 'nowrap',
-                      fontSize: '12px',
-                      fontWeight: '500',
-                    }}
-                    rel="noreferrer"
-                  >
-                    {lang.strings.buttons.earnRewards}
-                  </a>
-                </Button>
-              )}
-            {selectedAccount?.familyId === coinFamiliesMap.canton &&
               selectedAccount.type !== AccountTypeMap.subAccount &&
-              !selectedAccount.extraData?.isTransferPreApprovalEnabled && (
+              !selectedAccount.extraData?.isTransferPreApprovalEnabled &&
+              !!cantonAuthTokens?.refreshToken &&
+              !cantonUnauthorizedSyncError && (
                 <Button
                   variant="primary"
                   onClick={() =>
@@ -144,7 +137,9 @@ export const AccountPage: FC = () => {
               )}
             {selectedAccount?.familyId === coinFamiliesMap.canton &&
               selectedAccount.type !== AccountTypeMap.subAccount &&
-              !selectedAccount.extraData?.isMergeDelegationEnabled && (
+              !selectedAccount.extraData?.isMergeDelegationEnabled &&
+              !!cantonAuthTokens?.refreshToken &&
+              !cantonUnauthorizedSyncError && (
                 <Button
                   variant="primary"
                   onClick={() =>
@@ -161,6 +156,18 @@ export const AccountPage: FC = () => {
                   align="center"
                 >
                   {lang.strings.buttons.mergeDelegation}
+                </Button>
+              )}
+            {selectedAccount?.familyId === coinFamiliesMap.canton &&
+              (!cantonAuthTokens?.refreshToken ||
+                cantonUnauthorizedSyncError) && (
+                <Button
+                  variant="primary"
+                  onClick={() => {
+                    dispatch(openSyncCantonAccountPromptDialog());
+                  }}
+                >
+                  {lang.strings.buttons.resync}
                 </Button>
               )}
             <Container pl={{ def: 0, mdlg: 4 }} gap={8}>