+++ packages/cysync-core/src/store/canton/index.ts
@@ -13,11 +13,15 @@ export interface ICantonAuthTokens {
 export interface ICantonState {
   cantonAuthTokens: ICantonAuthTokens | undefined;
   isLoaded: boolean;
+  hasUnauthorizedSyncError: boolean;
+  isSyncPrompted: boolean;
 }
 
 const initialState: ICantonState = {
   isLoaded: false,
   cantonAuthTokens: {},
+  hasUnauthorizedSyncError: false,
+  isSyncPrompted: false,
 } as ICantonState;
 
 export const cantonSlice = createSlice({
@@ -37,13 +41,37 @@ export const cantonSlice = createSlice({
       state.cantonAuthTokens = undefined;
       return state;
     },
+    setCantonUnauthorizedSyncError: (
+      state,
+      action: PayloadAction<{ hasError: boolean }>,
+    ) => {
+      state.hasUnauthorizedSyncError = action.payload.hasError;
+      return state;
+    },
+    setIsSyncPrompted: (
+      state,
+      action: PayloadAction<{ isPrompted: boolean }>,
+    ) => {
+      state.isSyncPrompted = action.payload.isPrompted;
+      return state;
+    },
   },
 });
 
-export const { updateCantonAuthTokens, clearCantonAuthTokens } =
-  cantonSlice.actions;
+export const {
+  updateCantonAuthTokens,
+  clearCantonAuthTokens,
+  setCantonUnauthorizedSyncError,
+  setIsSyncPrompted,
+} = cantonSlice.actions;
 
 export const selectCantonAuthTokens = (state: RootState) =>
   state.canton.cantonAuthTokens;
+
+export const selectCantonUnauthorizedSyncError = (state: RootState) =>
+  state.canton.hasUnauthorizedSyncError;
+
+export const selectIsSyncPrompted = (state: RootState) =>
+  state.canton.isSyncPrompted;
 
 export default cantonSlice.reducer;