+++ packages/cysync-core-constants/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cypherock/cysync-core-constants
 
+## 0.7.1
+
+### Patch Changes
+
+- [#856](https://github.com/Cypherock/cypherock-cysync/pull/856) [`d23870c6`](https://github.com/Cypherock/cypherock-cysync/commit/d23870c6261211735dffe9832317e48ca7677a55) Thanks [@AryanCypherock](https://github.com/AryanCypherock)! - login flow added in canton accounts
+
+- Updated dependencies [[`223a00aa`](https://github.com/Cypherock/cypherock-cysync/commit/223a00aa4cc8b06f75b75b58cbebb7dfcbddeb76)]:
+  - @cypherock/coins@0.7.0
+
+## 0.7.1-pre.0
+
+### Patch Changes
+
+- [#856](https://github.com/Cypherock/cypherock-cysync/pull/856) [`d23870c6`](https://github.com/Cypherock/cypherock-cysync/commit/d23870c6261211735dffe9832317e48ca7677a55) Thanks [@AryanCypherock](https://github.com/AryanCypherock)! - login flow added in canton accounts
+
+- Updated dependencies [[`223a00aa`](https://github.com/Cypherock/cypherock-cysync/commit/223a00aa4cc8b06f75b75b58cbebb7dfcbddeb76)]:
+  - @cypherock/coins@0.7.0-pre.0
+
 ## 0.7.0
 
 ### Minor Changes